* commit 'e58bd735cb9b3ec88a3e6a55be5e35037c0c7a06':
  Create the alarm after setting the time.

+++ src/com/android/deskclock/provider/ClockDatabaseHelper.java
@@ -137,7 +137,7 @@ class ClockDatabaseHelper extends SQLiteOpenHelper {
             // Check if we are trying to re-use an existing id.
             Object value = values.get(ClockContract.AlarmsColumns._ID);
             if (value != null) {
-                int id = (Integer) value;
+                long id = (Long) value;
                 if (id > -1) {
                     final Cursor cursor = db.query(ALARMS_TABLE_NAME,
                             new String[]{ClockContract.AlarmsColumns._ID},
